Device-specific local size computation
Enable devices to override the logic of how to compute the
local dimensions, when they are left to the runtime.

Adds an alternative local size computation option for SIMD machines with
small WG support that want to maximize the vector utilization primarily.
This is enabled for HSA devices with less than 256 sized work groups.

Also some minor cleanups.